refactor: move sync system into `sync/legacy`
