+++ dash-spv/src/sync/legacy/masternodes/manager.rs
